+++ CHANGELOG.md
@@ -39,7 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.1.1...HEAD
-[0.1.1]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.1.1
+[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.1.0
 [0.0.1]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.0.1